requirements.txt
pillow>=10.0.0
tkinter-tooltip>=2.0.0
pillow>=10.0.0
reportlab>=4.0.0
python-barcode>=0.15.0
qrcode>=7.4.0
